refactor（v0.2.1-optimal）：拆分 chat/messages codec 聚合测试
