+++ src-tauri/src/protocol/codec/tests/chat_messages/support.rs
@@ -1,0 +1,5 @@
+use crate::protocol::codec::error::UnsupportedFeatures;
+
+pub(super) fn reject_features(e: &UnsupportedFeatures) -> Vec<String> {
+    e.features.clone()
+}
